+++ src/app/dashboard/fichier/fichier-de-base/programme/programme-list/programme-list.component.ts
@@ -3,6 +3,8 @@ import { ProgrammeService } from 'src/app/shared/services/programme.service';
 import { Programme, ListProgrammeResponse } from 'src/app/models/programme.model';
 import { Router } from '@angular/router';
 import {DataService} from '../../../../../shared/services/data.service';
+import {DELETE_CONFIRMATION} from '../../../../../constants/urlConstants';
+import {el} from '@angular/platform-browser/testing/src/browser_util';
 
 @Component({
   selector: 'app-programme-list',
@@ -33,13 +35,16 @@ export class ProgrammeListComponent implements OnInit {
   }
 
   onDelete(id) {
-    this.programmeService.deleteProgramme(id).subscribe((res) => {
-        this.programmes = this.programmes.filter((action) => {
-          return action.identifiant !== id;
-        });
-        this.router.navigate(['/dashboard/fichier/base/programme']);
-      }
-    );
+    const response = confirm(DELETE_CONFIRMATION);
+   if (response) {
+     this.programmeService.deleteProgramme(id).subscribe((res) => {
+         this.programmes = this.programmes.filter((action) => {
+           return action.identifiant !== id;
+         });
+         this.router.navigate(['/dashboard/fichier/base/programme']);
+       }
+     );
+   }
   }
 
 }
